% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getd}
\alias{getd}
\title{Finds the double derivative of A}
\usage{
getd(u1, v, du1da, dvda)
}
\arguments{
\item{u1}{u1}

\item{v}{v}

\item{du1da}{First derivative of u1 with respect to parameter a}

\item{dvda}{First derivative of v with respect to parameter a}
}
\value{
d2logAda2
}
\description{
Finds the double derivative of A with with respect to 
a, (a, b), b , (a, b) in respective templates from right to left. 
This first derivative is evaluated at the optimal (a_hat, b_hat).
u1, v and u2 constitute the equations required for evaluating 
the first and second order derivatives of A with respect to parameters 
a and b
}
