% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getd2logAda2}
\alias{getd2logAda2}
\title{Finds the double derivative of A with with respect to a. This first derivative is evaluated at the optimal (a_hat, b_hat).}
\usage{
getd2logAda2(u1, v, du1da, dvda)
}
\arguments{
\item{u1}{u1}

\item{v}{v}

\item{du1da}{First derivative of u1 with respect to parameter a}

\item{dvda}{First derivative of v with respect to parameter a}
}
\value{
d2logAda2
}
\description{
u1, v and u2 constitute the equations required for evaluating the first and second order derivatives of A with respect to parameters a and b
}
