% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getd2logAdadb}
\alias{getd2logAdadb}
\title{Finds the double derivative of A with with respect to a and b. This first derivative is evaluated at the optimal (a_hat, b_hat).}
\usage{
getd2logAdadb(u2, v, du2da, dvda)
}
\arguments{
\item{u2}{u2}

\item{v}{v}

\item{du2da}{First derivative of u2 with respect to a}

\item{dvda}{First derivative of v with respect to a}
}
\value{
d2logAdadb
}
\description{
u1, v and u2 constitute the equations required for evaluating the first and second order derivatives of A with respect to parameters a and b
}
