% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getd2logAdb2}
\alias{getd2logAdb2}
\title{Finds the double derivative of A with with respect to b. This first derivative is evaluated at the optimal (a_hat, b_hat).}
\usage{
getd2logAdb2(u2, v, du2db, dvdb)
}
\arguments{
\item{u2}{u2}

\item{v}{v}

\item{du2db}{First derivative of u2 with respect to b}

\item{dvdb}{First derivative of v with respect to}
}
\value{
d2logAdb2
}
\description{
u1, v and u2 constitute the equations required for evaluating the first and second order derivatives of A with respect to parameters a and b
}
