% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getd2logAdbda}
\alias{getd2logAdbda}
\title{Finds the double derivative of A with with respect to a and b.
This first derivative is evaluated at the optimal (a_hat, b_hat).}
\usage{
getd2logAdbda(u1, v, du1db, dvdb)
}
\arguments{
\item{u1}{u1}

\item{v}{v}

\item{du1db}{First derivative of u1 with respect to b}

\item{dvdb}{First derivative of v with respect to b}
}
\value{
d2logAdbda
}
\description{
u1, v and u2 constitute the equations required for evaluating
the first and second order derivatives of A with respect to parameters 
a and b
}
