% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getDataStatistics}
\alias{getDataStatistics}
\title{Evaluates statistics of the read counts corresponding to the gene}
\usage{
getDataStatistics(sp, spOne, spTwo)
}
\arguments{
\item{sp}{The complete (normalized and filtered) read count data 
corresponding to the gene in question}

\item{spOne}{The (normalized and filtered) read count data corresponding 
to the first sub-population}

\item{spTwo}{The (normalized and filtered) read count data corresponding 
to the second sub-population}
}
\value{
geneStats A vector containing 6 values corresponding to the gene 
data(maximum, minimum, mean, standard deviation, upper multiple of standard
deviation and lower multiple of standard deviation)
}
\description{
Takes in the complete read count vector corresponding to the
gene (sp) and also the data corresponding to the two sub-populations 
(sp1 and sp2)
}
