% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getdvdb}
\alias{getdvdb}
\title{Finds the first derivative of v with respect to b. 
This first derivative is evaluated at the optimal (a_hat, b_hat).}
\usage{
getdvdb(coefficients, r)
}
\arguments{
\item{coefficients}{the optimal values of a and b}

\item{r}{the rank vector}
}
\value{
dvdb
}
\description{
u1, v and u2 constitute the equations required for evaluating
the first and second order derivatives of A with respect to parameters 
a and b
}
