% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getv}
\alias{getv}
\title{Computes v}
\usage{
getv(coefficients, r)
}
\arguments{
\item{coefficients}{the optimal values of a and b}

\item{r}{the rank vector}
}
\value{
v
}
\description{
u1, v and u2 constitute the equations required for evaluating the first and second order derivatives of A with respect to parameters a and b
}
