% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{TMMnormalization}
\alias{TMMnormalization}
\title{TMM Normalization.}
\usage{
TMMnormalization(countTable)
}
\arguments{
\item{countTable}{The filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}
}
\value{
countTableTMM
}
\description{
Trimmed Means of M values (TMM) normalization (on the basis of edgeR package)
}
