% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getI}
\alias{getI}
\title{Computes the Fisher Information Matrix}
\usage{
getI(results)
}
\arguments{
\item{results}{A vector containing 5 values (a, b, A, number of bins, R2)}
}
\value{
I  The Fisher Information Matrix
}
\description{
The Fisher Information Matrix and its derivatives are essential to evulate the minima of log likelihood
}
