% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{getu1}
\alias{getu1}
\title{Computes u1}
\usage{
getu1(coefficients, r, nbits)
}
\arguments{
\item{coefficients}{the optimal values of a and b}

\item{r}{the rank vector}

\item{nbits}{number of bits for mpfr function}
}
\value{
u1
}
\description{
u1, v and u2 constitute the equations required for evaluating the first and second order derivatives of A with respect to parameters a and b
}
