% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSeq.R
\name{ROSeq}
\alias{ROSeq}
\title{ROSeq - A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
ROSeq(countData, condition, numCores)
}
\arguments{
\item{countData}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{condition}{Labels for the two sub-populations}

\item{numCores}{The number of cores to be used}
}
\value{
pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
